Extract phoronix package downloader
Benchmarks now download packages when run by the o4bc-bench tool.

Fixes #11, fixes #12
