+++ phoronix_benchmark.json.template
@@ -2,7 +2,12 @@
     "name": "PUT_NAME_HERE",
     "description": "PUT_DESCRIPTION_HERE",
     "default_preset": "PUT_DEFAULT_PRESETS_HERE",
-    "setup_command": "./setup.sh",
+    "setup_command": [
+        "pip install progressbar requests",
+        "./phoronix_downloader.py",
+        "@INSTALLER_FILENAME@",
+        "./setup.sh"
+    ],
     "run_command": "PUT_RUN_COMMAND_HERE",
     "test_command": ["pip install flake8", "flake8 --exclude .venv"],
     "stats":
